feature/animation

+++ mandatory/src/core/game.c
@@ -6,7 +6,7 @@
 /*   By: eandre-f <eandre-f@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/01 20:07:51 by eandre-f          #+#    #+#             */
-/*   Updated: 2023/03/11 22:27:57 by eandre-f         ###   ########.fr       */
+/*   Updated: 2023/03/12 01:06:16 by eandre-f         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,7 +30,8 @@ int	game_setup(t_game *game)
 	save_canvas_background(game->canvas);
 	init_textures(game);
 	game->control.fov_ratio = tanf(FOV_RAD / (game->canvas->width - 1));
-	game->animation.all_perpend = ft_calloc(game->canvas->width, sizeof(int));
+	game->animation.all_perpend = ft_calloc(game->canvas->width,
+			sizeof(double));
 	game->animation.inverse_determinant = 1.0 / (game->player.plane.x
 			* game->player.dir.y - game->player.dir.x * game->player.plane.y);
 	return (0);
